* moved constants to shared folder
