+++ shared/src/packages/appeals/appeals.ts
@@ -2,5 +2,8 @@ export { MIN_DESCRIPTION_LENGTH,
     MAX_DESCRIPTION_LENGTH,
     MAX_NUMBER_OF_PHOTOS,
     MAX_ADDRESS_LENGTH,
-    MIN_ADDRESS_LENGTH
+    MIN_ADDRESS_LENGTH,
+    MAX_FULL_NAME_LENGTH,
+    MIN_FULL_NAME_LENGTH,
+    Categories
 } from './libs/constants/constants.js';
